# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AGA Health Foundation Telehealth Readiness Survey API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /system/prospectus:
    get:
      operationId: listSystemProspectuses
      tags: [system-prospectus]
      summary: List all prospectuses (system admin)
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProspectusStatus"
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusListResponse"
  /system/prospectus/{id}:
    get:
      operationId: getSystemProspectus
      tags: [system-prospectus]
      summary: Prospectus detail with reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusDetailResponse"
  /system/prospectus/{id}/reviews:
    post:
      operationId: addProspectusReview
      tags: [system-prospectus]
      summary: Add review comment or request revision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectusReviewInput"
      responses:
        "201":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusDetailResponse"
  /system/prospectus/{id}/approve:
    post:
      operationId: recordProspectusApproval
      tags: [system-prospectus]
      summary: Record dual-approval decision (research_leadership or platform_ops)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectusApprovalInput"
      responses:
        "200":
          description: Approval recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusDetailResponse"
  /system/prospectus/{id}/provision-study:
    post:
      operationId: provisionStudyFromProspectus
      tags: [system-prospectus]
      summary: Create draft studies row after dual approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionStudyFromProspectusInput"
      responses:
        "201":
          description: Study provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionStudyResult"
components:
  schemas:
...
    ProspectusStatus:
      type: string
      enum:
        - draft
        - submitted
        - under_review
        - revision_requested
        - approved
        - rejected
        - withdrawn
    ProspectusApprovalRole:
      type: string
      enum:
        - research_leadership
        - platform_ops
...
    ProspectusApprovalSlot:
      type: object
      required: [role, decision, comments, decidedAt]
      properties:
        role:
          $ref: "#/components/schemas/ProspectusApprovalRole"
        decision:
          type: ["string", "null"]
          enum: [approved, rejected, null]
        comments:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
          format: date-time
    ProspectusPublic:
      type: object
      required: [publicId, status, submitterEmail, submitterName]
      properties:
        id:
          type: integer
        publicId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProspectusStatus"
        submitterEmail:
          type: string
        submitterName:
          type: string
        title:
          type: string
        principalInvestigator:
          type: string
        proposedSlug:
          type: ["string", "null"]
        studyTemplate:
          type: string
        linkedStudySlug:
          type: ["string", "null"]
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ProspectusApprovalSlot"
        attachmentCount:
          type: integer
    ProspectusListItem:
      type: object
      properties:
        id:
          type: integer
        publicId:
          type: string
        status:
          $ref: "#/components/schemas/ProspectusStatus"
        title:
          type: string
        submitterName:
          type: string
        submitterEmail:
          type: string
        proposedSlug:
          type: ["string", "null"]
        linkedStudySlug:
          type: ["string", "null"]
    ProspectusListResponse:
      type: object
      required: [prospectuses]
      properties:
        prospectuses:
          type: array
          items:
            $ref: "#/components/schemas/ProspectusListItem"
    ProspectusDetailResponse:
      type: object
      required: [prospectus, reviews]
      properties:
        prospectus:
          $ref: "#/components/schemas/ProspectusPublic"
        reviews:
          type: array
          items:
            type: object
    ProspectusReviewInput:
      type: object
      required: [decision, comments]
      properties:
        decision:
          type: string
          enum: [comment, revision_requested]
        comments:
          type: string
        isInternal:
          type: boolean
    ProspectusApprovalInput:
      type: object
      required: [approvalRole, decision]
      properties:
        approvalRole:
          $ref: "#/components/schemas/ProspectusApprovalRole"
        decision:
          type: string
          enum: [approved, rejected]
        comments:
          type: string
    ProvisionStudyFromProspectusInput:
      type: object
      properties:
        slug:
          type: string
        responsesTable:
          type: string
        shortTitle:
          type: string
    ProvisionStudyResult:
      type: object
      properties:
        slug:
          type: string
        prospectusId:
          type: integer
        message:
          type: string
...